requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.115.4
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
h11==0.14.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.24.3
onnx==1.17.0
onnxruntime==1.19.2
packaging==24.1
pillow==11.0.0
protobuf==5.28.3
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.17
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
sympy==1.13.3
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
